refactor(cms): split editor normalizers
Move owner-editor page keys, normalization, and validation helpers into cms/editor.ts so cms/index.ts keeps the D1 fetch/query boundary and public re-exports.

Verified with pnpm --filter @anipotts/lib typecheck; pnpm --filter @anipotts/lib test; pnpm --filter @anipotts/lib build; pnpm dlx knip --no-exit-code; pnpm exec prettier --check packages/lib/src/cms/index.ts packages/lib/src/cms/editor.ts; git diff --check.
